requirements.txt
python-telegram-bot
schedule
transformers
torch
requests